Dev environment update
Dev environment has been updated; run "npm prune" to clean up old dependencies. Start the server with the command "server/npm run serve" (it will watch for changes on the server code) and rebuild the client using "npm run build"

I set up a command to copy the bundle.js file to the server directory and watch for changes so builds don't need to be done automatically - "npm run watch" however this is buggy as the copy watch function doesn't wait for webpack to finish compiling.