Changes:
user list is downloading, when a user joins a channel their object is sent to users in the channel, actions and reducers are updated to suit, selector for visible users which also performs sorting is created and working.
To do: users are not removed from the list when they leave. Switching between channels does not work properly when joining a channel, refreshing the user list when switching channels does not work properly